// apps/wallet-mobile/src/yoroi-wallets/balance/adapters/api/types.ts
export type ApiTokenId = `${string}.${string}`

export type ApiFutureTokenRecords = {
  [tokenId: string]: ApiFutureToken
}
export type ApiFutureToken = {
  supply: ApiTokenSupplyRecord
  offChain: ApiOffChainMetadataRecord
  onChain: ApiOnChainMetadataRecord
}

export type ApiOffChainMetadataRequest = ReadonlyArray<ApiTokenId>
export type ApiOffChainMetadataResponse = {[tokenId: ApiTokenId]: ApiOffChainMetadataRecord}
export type ApiOffChainMetadataRecord =
  | {
      tokenRegistry: Record<string, unknown> | undefined
      isValid: false
    }
  | {
      tokenRegistry: ApiTokenRegistryEntry
      isValid: true
    }

// CIP26 Token Registry
// https://github.com/cardano-foundation/cardano-token-registry#semantic-content-of-registry-entries
export type ApiTokenRegistryEntry = {
  subject: string
  name: ApiTokenRegistryProperty<string> // 1-50 bytes HR (specs say min 1, but there are entries with empty name)

  description?: ApiTokenRegistryProperty<string> // 0-500 bytes HR
  policy?: string // 56-120 bytes base16 used as preimage to produce a policyId
  logo?: ApiTokenRegistryProperty<string> // 0-87400 bytes base64 png i.e: `data:image/png;base64,${logo}`
  ticker?: ApiTokenRegistryProperty<string> // 2-9 bytes HR (specs say min 2, but there are entries with empty ticker)
  url?: ApiTokenRegistryProperty<string> // 0-250 bytes URI
  decimals?: ApiTokenRegistryProperty<number> // int 0-19
}

type ApiTokenRegistrySignature = {
  publicKey: string
  signature: string
}

export type ApiTokenRegistryProperty<T> = {
  signatures: Array<ApiTokenRegistrySignature>
  sequenceNumber: number
  value?: T
}